+++ optim/expression.go
@@ -26,7 +26,7 @@ type Expression interface {
 	IDs() []uint64
 
 	// Dims returns a slice describing the true dimensions of a given expression (scalar, vector, or matrix)
-	Dims() []uint64
+	Dims() []uint
 
 	//// Plus adds the current expression to another and returns the resulting
 	//// expression
